requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
dirtyjson==1.0.7
Django==3.2.11
django-environ==0.8.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==12.0.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
zope.interface==5.4.0
